+++ api/web/src/pages/connection/main.ts
@@ -42,6 +42,22 @@ const router = VueRouter.createRouter({
                 name: `layer-incoming-schema`,
                 component: () => import('../../components/ETL/Layer/LayerIncomingSchema.vue')
             },{
+                path: 'incoming/mapping',
+                name: `layer-incoming-mapping`,
+                component: () => import('../../components/ETL/Layer/LayerIncomingMapping.vue')
+            },{
+                path: 'incoming/mapping/:schema',
+                name: `layer-incoming-mapping-schema`,
+                component: () => import('../../components/ETL/Layer/LayerIncomingMappingSchema.vue')
+            },{
+                path: 'incoming/mapping/:schema/query/new',
+                name: `layer-incoming-mapping-query-new`,
+                component: () => import('../../components/ETL/Layer/LayerIncomingMappingQuery.vue')
+            },{
+                path: 'incoming/mapping/:schema/query/:query',
+                name: `layer-incoming-mapping-query`,
+                component: () => import('../../components/ETL/Layer/LayerIncomingMappingQuery.vue')
+            },{
                 path: 'incoming/styles',
                 name: `layer-incoming-styles`,
                 component: () => import('../../components/ETL/Layer/LayerIncomingStyles.vue')